+++ patzilla-ui/navigator/app/document/document-view.js
@@ -15,7 +15,7 @@ OpsExchangeDocumentView = Backbone.Marionette.Layout.extend({
 
     // Namespace template variables to "data", also accounting for "templateHelpers".
     serializeData: function() {
-        var data = this.model.attributes;
+        var data = _.clone(this.model.attributes);
         _.extend(data, this.templateHelpers());
         return {data: data};
     },